+++ stable_baselines/common/tf_layers.py
@@ -42,6 +42,25 @@ def ortho_init(scale=1.0):
     return _ortho_init
 
 
+def mlp(input_tensor, layers, activ_fn=tf.nn.relu, layer_norm=False):
+    """
+    Create a multi-layer fully connected neural network.
+
+    :param input_tensor: (tf.placeholder)
+    :param layers: ([int]) Network architecture
+    :param activ_fn: (tf.function) Activation function
+    :param layer_norm: (bool) Whether to apply layer normalization or not
+    :return: (tf.Tensor)
+    """
+    output = input_tensor
+    for i, layer_size in enumerate(layers):
+        output = tf.layers.dense(output, layer_size, name='fc' + str(i))
+        if layer_norm:
+            output = tf.contrib.layers.layer_norm(output, center=True, scale=True)
+        output = activ_fn(output)
+    return output
+
+
 def conv(input_tensor, scope, *, n_filters, filter_size, stride,
          pad='VALID', init_scale=1.0, data_format='NHWC', one_dim_bias=False):
     """
