+++ stable_baselines/common/tf_util.py
@@ -173,29 +173,6 @@ def q_explained_variance(q_pred, q_true):
 
 
 # ================================================================
-# TF network blocks
-# ================================================================
-
-def mlp(input_ph, layers, activ_fn=tf.nn.relu, layer_norm=False):
-    """
-    Create a multi-layer fully connected neural network.
-
-    :param input_ph: (tf.placeholder)
-    :param layers: ([int]) Network architecture
-    :param activ_fn: (tf.function) Activation function
-    :param layer_norm: (bool) Whether to apply layer normalization or not
-    :return: (tf.Tensor)
-    """
-    output = input_ph
-    for i, layer_size in enumerate(layers):
-        output = tf.layers.dense(output, layer_size, name='fc' + str(i))
-        if layer_norm:
-            output = tf.contrib.layers.layer_norm(output, center=True, scale=True)
-        output = activ_fn(output)
-    return output
-
-
-# ================================================================
 # Global session
 # ================================================================
